Optimize code sample of QObject::isSignalConnected
Since isSignalConnected is meant to be used in performance critical
code, and that QMetaMethod::fromSignal is relatively slow, it is better
to have it a a static variable in the code sample, as a good
recommendation on how to use it.

Since QMetaMethod::fromSignal should not change during the life of the
program, it should be safe.  Also, if different threads run at the same
time, both should lead to the same result.

Change-Id: Ib6113d11ca93f216bc3a92aea4eaa4da6a4151ca
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>