Doc: Fixed QDoc's navigation mechanism.
The HTML pages need to display minimal navigation such
as a link to the module landing page or Qt 5's index.html.

-new variables available in qdocconf files:
    navigation.landingpage
    navigation.homepage
    navigation.cppclassespage
    navigation.qmltypespage
    buildversion

The navigation variables are for the new navigation bar.
The buildversion variable is to insert documentation build
information into the pages.

Note about compatibility with Qt Creator docs:
The HTML template files for Qt Creator is retained to keep
Qt Creator compatible with 5.0.x and 5.1.x.

Change-Id: Ibb4d7cada19644204457d822b6b77b2aa7b70f8d
Reviewed-by: Martin Smith <martin.smith@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>